Sprint 4 (#38)
* Centrado el menu de amigos / pagos conjuntos.

* Creada la activity del formulario de pagos conjuntos.

* Código para fijar el tamaño de la imágen.

* Inicio del form para la creación de items de pago.

* Avances en la lista de usuarios

* Bug Checkbox

No consigo hacer que se vea la checkbox al correr la app, pero está en el layout

* Checkbox arreglada

Ahora ya se ve y funciona

* Lógica de selección de usuarios con checkbox

* Updates en la lógica de cambio de pantalla

* Lanzamiento de actividad desde fragment

* Avances en la validación del formulario

* Bug raro al añadir el pago creado.

* Encontrado donde da el error.

* Avances en la solución de errores

* Añadidas las imagenes como URIs a el Pago Conjunto

* Small fix de Snackbar

* Pequeñas adiciones al PagoConjunto

* Añadida la fecha límite al pago conjunto

* Validada la fecha límite.

* Inicio del form para la creación de items de pago.

* Parcelable esta molestando.

* Parcelable creo que esta arreglado.

* Añadida la fecha al pago conjunto y mostrada en el layout.

* Bug fix de serialización

* Ahora la imagen del pago se puede ver en la lista.

* Formulario para crear items de pago.

* Formulario para crear items de pago.

* Persistencia de imagenes en pagos conjuntos.

* Arreglado el problema de imagen default

* Fix codigo deprecado

* Filtrar pagos por usuario

Faltaba añadir una manera de que cada usuario solo pudiese ver sus pagos en vez de todos los existentes.

* Cosas de la base de datos.

* Items Pago Conjunto se guardan en la base de datos y se recuperan.

* Amigos obtenidos desde base de datos e imagenes de amigos cargadas.

* Lógica para añadir amigos a pagos

* Logs borrados

* Solicitudes amigos and menu added (#33)

* Mockup mas o menos implementado

* Branch change

* Menu not working

* Menu funcionando mas o menos

* Menu working y refactos a algunas vistas

* Refactor pequeño y adapter de solicitudes entrantes 'hecho'

* Lista de solicitudes aparece (falta aceptar rechazar), lista de envio sin funcionalidad

* Loading friends in friend list and correct requests

* Friend reqs working

* Refactor view and confirm on remove friend requested

* Added no info view to grups

* BugFix could send request to people that were already friend

* BugFix on friendrec could be send and accept one of the same user, leaving with a request of a user that is already a friend

* Merge with develop

---------


* Color changes

* Keyboard se esconde al buscar amigos

* Color change

* Merge y damian constraint

* With bugs

* Todo persiste.

* No hay usuarios que pagan

* Amigos.

* Fix friendRequests

* BugFix

* Item linea recycler fix

* Imagenes de usuarios

---------

+++ app/src/main/java/com/moneyguardian/ui/ItemPagosFragment.java
@@ -11,13 +11,12 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
-import com.moneyguardian.ListaBalanceItemAdapter;
+import com.moneyguardian.adapters.ListaBalanceItemAdapter;
 import com.moneyguardian.R;
 import com.moneyguardian.modelo.Usuario;
 import com.moneyguardian.modelo.UsuarioParaParcelable;
 
 import java.util.HashMap;
-import java.util.Map;
 
 public class ItemPagosFragment extends Fragment {
 
@@ -25,13 +24,13 @@ public class ItemPagosFragment extends Fragment {
     private static final String USERS_AND_PAYMENTS = "balance";
 
     private String nombreItem;
-    private HashMap<UsuarioParaParcelable,Integer> mapBalance = new HashMap<>();
+    private HashMap<UsuarioParaParcelable,Double> mapBalance = new HashMap<>();
 
     private TextView tvNombreItem;
     private RecyclerView rvBalance;
 
     public static ItemPagosFragment newInstance(String param1,
-                                                HashMap<UsuarioParaParcelable,Integer> param3) {
+                                                HashMap<UsuarioParaParcelable,Double> param3) {
         ItemPagosFragment fragment = new ItemPagosFragment();
         Bundle args = new Bundle();
         args.putString(NAME, param1);
@@ -46,7 +45,7 @@ public class ItemPagosFragment extends Fragment {
         super.onCreate(savedInstanceState);
         if (getArguments() != null) {
             nombreItem = getArguments().getString(NAME);
-            mapBalance = (HashMap<UsuarioParaParcelable, Integer>) getArguments().get(USERS_AND_PAYMENTS);
+            mapBalance = (HashMap<UsuarioParaParcelable, Double>) getArguments().get(USERS_AND_PAYMENTS);
         }
     }
